Refactor GolemService to support network stats
